refactor(issue-detail): extract shared MarkdownEditor component
Deduplicates the inline markdown editors in IssueDetailActionBar and
IssueDetailCommentList into a single MarkdownEditor under
frontend/src/react/components/MarkdownEditor/, mirroring the Vue
MarkdownEditor folder layout. Adds mode/transform/maxHeight props so
the comment list can supply its #issue-ref linker and the read-only
view path. Tightens the prop type via a discriminated union so editor
mode requires onChange (previous loose typing silently no-op'd
keystrokes).
