+++ frontend/src/react/components/MarkdownEditor/index.ts
@@ -1,0 +1,1 @@
+export { MarkdownEditor } from "./MarkdownEditor";
