+++ frontend/tsconfig.app.json
@@ -10,6 +10,7 @@
   "exclude": [
     "src/**/__tests__/*",
     "src/react/**/*.tsx",
+    "src/react/components/MarkdownEditor/index.ts",
     "src/react/components/task-run-log/index.ts",
     "src/react/components/monaco/index.ts",
     "src/react/components/revision/index.ts"
