+++ frontend/tsconfig.vitest.json
@@ -9,6 +9,7 @@
   ],
   "exclude": [
     "src/react/**/*.tsx",
+    "src/react/components/MarkdownEditor/index.ts",
     "src/react/components/task-run-log/index.ts",
     "src/react/components/monaco/index.ts",
     "src/react/components/revision/index.ts"
